Add ViewModel for the Recipe screen
Adds a view model to the recipe screen to move logic outside of the view class. Previously, any logic that had dealt with any direct input from the user would.
Slight modification to the table view in `RecipeScreen` to support longer lists of ingredients and steps, as the current version would cut off information if too many items were present.