Add autocomplete to ingredient name
Replaces the default `TextBox` for the ingredient name input with a custom `AutocompleteTextBox` control, allowing for ingredient names to be filtered and completed automatically.

An `AutocompleteTextBox` is used instead of a standard `TextBox` as it allows autocomplete for words that contain the typed text, unlike `TextBox` which only autocompletes words that begin with the text.