Update class diagram
Updates the class diagram to include changes to the system made in sprint 2.
Does not update all parts of the sprint 1 items, as I would like to ask Corley what he would like us to do for the shared resources. I want to avoid a thousand arrows going around the screen, if possible.