Updates view model for RecipesListScreen
Updates `RecipeListViewModel` to incorporate databinding, moving much of the application logic away from the screen for easier testing. Dur to how most panels work in Windows Forms, databinding to them isn't possible out of the box. Some logic had to remain in the screen to allow the lists to populate.