Add tag list to recipe information
Adds the list of tags to both the `RecipeListItem` and the `RecipeScreen`. This gives the user a better idea of what the recipe is going to be like.

On a more practical level, it allows us to get better feedback regarding whether our filtering based on tags is working properly.