Update font to be Calibri
Updates all fonts in the application to use `Calibri` instead of `Segoe UI`. Some control sizes and position had to be adjusted slightly as `Calibri` is slightly smaller than `Segoe UI`, but the overall look and layout is almost identical to the previous version.

This was done to keep a consistent appearance between UI elements, as different elements on the screen would have different font families, seemingly at random.