Update the class diagram
Updates the class diagram, adding new classes and members.

Splits the diagram into three pages, one for the desktop client, the web client, and for the server. This means that shared resources will need to be updated on all three pages separately, but it allows us to organize each section much more effectively.