Add max length constraints to text boxes
Adds constraints to the maximum length of text boxes to match the maximum length of attributes in the database.
Password text boxes have been given a constraint to match the maximum password length of 20.

This should prevent errors due to users inputting text that is entirely too long. While the server has input validation as well, this serves as the first layer of protection.