Replace dialog windows with components
Replaces all dialog window classes with user components, allowing them to be displayed within the same window as the application itself.

This requires the code to display the dialog to be updated. Instead of showing the dialog in an if statement and having the logic within the body, the logic must be handled within the `DialogClosed` event handler. Exceptions cannot be caught as they were previously. Now, they are stored as a parameter and can be accessed within the event handler.