Add message when lists are empty
Adds a message for when the ingredient or recipe list is empty, informing the user that there was nothing to be displayed.

This reduces confusion for the user when nothing is shown, as it confirms that the operation was completed. Otherwise, the user may think that the operation is still in progress.