Remove conversion to UTC
Removes conversion of the `dateTimeToGenerateFrom` parameter for `GenerateDateTimesFromWeekToNextWeek`, replacing it with an assignment that removes the time portion of the `DateTime`.

This allows the method to generate two-week long spans of `DateTime`s relative to the raw date of the `DateTime` passed in, regardless of what the date would be in UTC. If it needs to be done in UTC, the conversion can be done outside the method first.