Implement getting missing ingredients, cook recipe
Adds the option to get the active user's missing ingredients for a specified recipe and the ability to remove all of a recipe's ingredient's from the active user's pantry.

Both methods will be used when the user is on a recipe screen, allowing them to mark a recipe as completed. If the user doesn't have enough ingredients to complete it, they will be informed about which ingredients they are missing.