Add button to cook recipe
Adds a button on `RecipeScreen` to complete a recipe. When completed, all of the recipe's ingredients will be removed from the active user's pantry. If they do not have the ingredients to complete the recipe, they will be informed about which ingredients they are missing.

This allows the user to quickly manage the ingredients they used to complete a given recipe.