Adjust registration screen UI
Adjusts the UI on `RegistrationScreen`, adding a title to the top of the screen, adjusting the placement of the user input, and resizing the heights of the rows on the top and bottom of the screen.

This makes the screen feel a little less barren and reduces the chances that the UI will overflow on certain screens.