Allow duplicate planned recipes in meals
Allows the same recipe to be added multiple times to the planned meals for a given day and meal.

The database had to be updated to allow this, removing the `UNIQUE` constraint from the `PlannedMeals` table. The server now sends the planned meal's unique id with the other information, and the clients send back the id when removing the meal.